+++ lib/go-atscfg/atscfg.go
@@ -31,6 +31,8 @@ import (
 // TODO move in config gen PR
 const InvalidID = -1
 
+const DefaultATSVersion = "5" // TODO Emulates Perl; change to 6? ATC no longer officially supports ATS 5.
+
 const HeaderCommentDateFormat = "Mon Jan 2 15:04:05 MST 2006"
 
 type ServerInfo struct {
@@ -39,6 +41,7 @@ type ServerInfo struct {
 	CDNID                         int
 	DomainName                    string
 	HostName                      string
+	HTTPSPort                     int
 	ID                            int
 	IP                            string
 	ParentCacheGroupID            int